最后如果不存在则不用删除transferID的目录，scow-sync可以实现前台执行

+++ scowsync.py
@@ -118,5 +118,6 @@ class ScowSync:
                 self.thread_pool.submit(
                     self.__transfer_file, entity_file.subpath)
         self.thread_pool.shutdown()
-        os.rmdir(os.path.join(self.base_path, str(self.transfer_id)))
+        if os.path.exists(os.path.join(self.base_path, str(self.transfer_id))):
+            os.rmdir(os.path.join(self.base_path, str(self.transfer_id)))
         sys.exit()
